refactor: extract sign-up form logic into a composable for reusability and testability
