+++ src/components/SignUpForm/SignUpForm.vue
@@ -1,68 +1,10 @@
 <script setup lang="ts">
-import { SignUpUsecase } from '~/features/auth/application/signup/SignUpUsecase';
 import SignUpFormPresentation from './SignUpFormPresentation.vue';
-import { isErrorResponse } from '~/features/shared/domain/Response';
-import type { SignUpFormState, SignUpFormValues } from './SignUpForm.types';
 import { LocalStorageAuthRepository } from '~/features/auth/infrastructure/LocalStorageAuthRepository';
+import useSignUpForm from './composables/useSignUpForm';
 
 const repository = new LocalStorageAuthRepository();
-const useCase = new SignUpUsecase(repository);
-const defaultSignUpFormState: SignUpFormState = {
-  isLoading: false,
-  successMessage: '',
-  values: {
-    email: '',
-    password: '',
-    withInfo: false,
-  },
-  errors: {
-    email: null,
-    password: null,
-    global: null,
-  },
-};
-const formState = ref<SignUpFormState>(defaultSignUpFormState);
-
-const handleChange = (field: keyof SignUpFormValues, value: string) => {
-  if (field === 'withInfo') {
-    formState.value.values.withInfo = value === 'true';
-  } else {
-    formState.value.values[field] = value;
-  }
-};
-
-const handleSubmit = async () => {
-  formState.value.isLoading = true;
-  formState.value.errors.global = '';
-  formState.value.successMessage = '';
-
-  const response = await useCase.execute(
-    formState.value.values.email,
-    formState.value.values.password,
-    formState.value.values.withInfo
-  );
-
-  if (isErrorResponse(response)) {
-    formState.value.errors = {
-      email: response.error.email || null,
-      password: response.error.password || null,
-      global: response.error.global || null,
-    };
-    formState.value.successMessage = defaultSignUpFormState.successMessage;
-    formState.value.isLoading = false;
-
-    return;
-  }
-
-  formState.value.values = response.value.values;
-  formState.value.successMessage = response.value.successMessage;
-  formState.value.errors = {
-    email: null,
-    password: null,
-    global: null,
-  };
-  formState.value.isLoading = false;
-};
+const { formState, handleChange, handleSubmit } = useSignUpForm(repository)();
 </script>
 
 <template>
